+++ main/ad2_cli_cmd.h
@@ -24,13 +24,14 @@
 #ifndef _AD2_CLI_CMD_H_
 #define _AD2_CLI_CMD_H_
 
-#define AD2_REBOOT   "reboot"
+#define AD2_REBOOT   "restart"
 #define AD2_NETMODE  "netmode"
 #define AD2_BUTTON   "button"
 #define AD2_CODE     "code"
 #define AD2_VPADDR   "vpaddr"
 #define AD2_SOURCE   "ad2source"
 #define AD2_TERM     "ad2term"
+#define AD2_LOGMODE  "logmode"
 #define AD2_UNLOCKER "teaser :c)"
 
 #ifdef __cplusplus