+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] WIP
+- [X] CORE: Improve: Command <strike>reboot</strike> -> restart. Much better, almost did a reboot on my host :)
+- [X] CORE: Tidy: Command help and moved to \r\n on all HOST communication same as AD2*
+- [X] CORE: New: Command: logmode {debug|info|none}. Set logging mode to get more | less details.
+- [X] CORE: Improve: Moved esp host UART init to hal_.
 - [X] CORE: Improve: New formatting utils to make it easy to build messages to send and route all host com through a one api to allow for management <strike>printf</strike>.
 - [X] CORE: Improve: Refactor init sequence and error checking to get the network up before connecting to the AD2*.
 - [X] STSDK: Improve: Set default 'stenable' to Y. At first boot it starts and initializes the st nvme partition allowing stserial etc to be called. It is still necessary to add the stserial etc before it will show as an AC on WiFi scanners. From the factory with Keys it will just show up at first boot.